Reconnection, zero lenght receive, timeout...
Not the best but tested working. It now reconnects in case of network loss and in case of no NMEA sent to server for long and hence 0 bytes from server will also attempt reconnection.  Avoids too fast reconnections and also checks if the corrections keeps coming, otherwise also attempt a reconnection.